feat(agent): add lifecycle agent framework and reorganize agents
Introduce a reusable lifecycle agent framework under common/lifecycleagent
and the first contract definition lifecycle agent. Move existing
orchestration agents under agent/orchestration/ and add NATS stream helpers.
