--- agent/ih/launcher/launcher.go
@@ -1,85 +1,0 @@
-//  Copyright (c) 2025 Metaform Systems, Inc
-//
-//  This program and the accompanying materials are made available under the
-//  terms of the Apache License, Version 2.0 which is available at
-//  https://www.apache.org/licenses/LICENSE-2.0
-//
-//  SPDX-License-Identifier: Apache-2.0
-//
-//  Contributors:
-//       Metaform Systems, Inc. - initial API and implementation
-//
-
-package launcher
-
-import (
-	"net/http"
-
-	"github.com/eclipse-cfm/cfm/agent/common/identityhub"
-	"github.com/eclipse-cfm/cfm/agent/ih/activity"
-	"github.com/eclipse-cfm/cfm/assembly/httpclient"
-	"github.com/eclipse-cfm/cfm/assembly/serviceapi"
-	"github.com/eclipse-cfm/cfm/common/runtime"
-	"github.com/eclipse-cfm/cfm/common/system"
-	"github.com/eclipse-cfm/cfm/common/tokenexchange"
-	"github.com/eclipse-cfm/cfm/pmanager/api"
-	"github.com/eclipse-cfm/cfm/pmanager/natsagent"
-)
-
-const (
-	ActivityType                       = "identityhub-activity"
-	urlKey                             = "vault.url"
-	identityHubURLKey                  = "identityhub.url"
-	identityHubCredentialServiceURLKey = "identityhub.cs.url"
-	controlPlaneProtocolURLKey         = "controlplane.protocol.url"
-	controlPlaneDataServiceURLKey      = "controlplane.dataservice.url"
-	tokenExchangeURLKey                = "tokenexchange.url"
-	tokenFilePathKey                   = "tokenexchange.tokenFilePath"
-	audienceKey                        = "tokenexchange.audience"
-)
-
-func LaunchAndWaitSignal(shutdown <-chan struct{}) {
-	config := natsagent.LauncherConfig{
-		AgentName:    "IdentityHub Agent",
-		ServiceName:  "cfm.agent.identityhub",
-		ConfigPrefix: "ihagent",
-		ActivityType: ActivityType,
-		AssemblyProvider: func() []system.ServiceAssembly {
-			return []system.ServiceAssembly{
-				&httpclient.HttpClientServiceAssembly{},
-			}
-		},
-		NewProcessor: func(ctx *natsagent.AgentContext) api.ActivityProcessor {
-			httpClient := ctx.Registry.Resolve(serviceapi.HttpClientKey).(http.Client)
-			ihURL := ctx.Config.GetString(identityHubURLKey)
-			vaultURL := ctx.Config.GetString(urlKey)
-			ihCsURL := ctx.Config.GetString(identityHubCredentialServiceURLKey)
-			cpProtocolURL := ctx.Config.GetString(controlPlaneProtocolURLKey)
-			cpDataServiceURL := ctx.Config.GetString(controlPlaneDataServiceURLKey)
-
-			if err := runtime.CheckRequiredParams(identityHubURLKey, ihURL); err != nil {
-				panic(err)
-			}
-
-			provider := tokenexchange.NewTokenExchangeProvider(ctx.Config.GetString(tokenFilePathKey),
-				tokenexchange.WithTokenExchangeUrl(ctx.Config.GetString(tokenExchangeURLKey)),
-				tokenexchange.WithTokenExchangeAudience(ctx.Config.GetString(audienceKey)),
-				tokenexchange.WithHttpClient(&httpClient))
-
-			return activity.NewProcessor(&activity.Config{
-				Client:      &httpClient,
-				LogMonitor:  ctx.Monitor,
-				IdentityAPIClient: identityhub.HttpIdentityAPIClient{
-					BaseURL:       ihURL,
-					TokenProvider: provider,
-					HttpClient:    &httpClient,
-				},
-				VaultURL:             vaultURL,
-				CredentialServiceURL: ihCsURL,
-				ProtocolServiceURL:   cpProtocolURL,
-				DataServiceURL:       cpDataServiceURL,
-			})
-		},
-	}
-	natsagent.LaunchAgent(shutdown, config)
-}
