+++ agent/orchestration/onboarding/types.go
