+++ agent/orchestration/registration/activity/activity_test.go
