+++ docker/Dockerfile.edcvagent.dockerfile
@@ -23,7 +23,7 @@ COPY . .
 
 # Build the agent binary
 RUN CGO_ENABLED=0 GOOS=$TARGETOS GOARCH=$TARGETARCH \
-    go build -ldflags="-s -w" -o bin/edcvagent ./agent/edcv/cmd/server/main.go
+    go build -ldflags="-s -w" -o bin/edcvagent ./agent/orchestration/edcv/cmd/server/main.go
 
 # Production stage
 FROM gcr.io/distroless/static-debian12:nonroot
